add to the BL the add functions

+++ סמסטר א/C#/dotNet5782_9080_0477/DAL/Customer.cs
@@ -10,7 +10,7 @@ namespace IDAL
     {
         public struct Customer
         {
-            public int ID { get; set; }
+            public ulong  ID { get; set; }
             public string Name { get; set; }
             public string Phone { get; set; }
             public double Longitude { get; set; }
