+++ סמסטר א/c#/dotnet5782_9080_0477/DAL/DalObjectCustomer.cs
@@ -23,7 +23,7 @@ namespace DalObject
             return customers;
         }
 
-        public Customer GetSpecificCustomer(int id)
+        public Customer GetSpecificCustomer(ulong id)
         {
             try
             {
